requirements.txt
anthropic>=0.18.0
apscheduler>=3.10.0
asyncpg>=0.27.0
discord.py>=2.0.0
psycopg2-binary>=2.9.0
PyGithub>=1.58.0
python-dotenv>=0.21.0
requests>=2.31.0
SQLAlchemy>=2.0.0